[Refactor] 패키지명 모델로 변경
