+++ lodge/src/main/java/com/haot/lodge/domain/model/Lodge.java
@@ -1,4 +1,4 @@
-package com.haot.lodge.domain.entity;
+package com.haot.lodge.domain.model;
 
 
 import jakarta.persistence.Column;
